提取解析逻辑，增加最外层的容器
