+++ component/mini-parser/render/index.json
@@ -1,6 +1,6 @@
 {
   "component": true,
   "usingComponents": {
-    "render": "./render/index"
+    "render": "./index"
   }
 }
