+++ component/mini-parser/render/index.wxss
@@ -1,0 +1,13 @@
+.mini-parser-inline{
+  display: inline;
+}
+.mini-parser-inline-block{
+  display: inline-block;
+}
+.mini-parser-flex{
+  display: flex;
+}
+.mini-parser-origin-strong{
+  font-weight: bold;
+}
+
